# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Jobseek dashboard API
  version: 0.1.0
  description: Versioned private dashboard API. All data routes require authentication.
servers:
  - url: /api/v1
paths:
...
  /opportunities/{id}/checks:
    parameters: [{ $ref: '#/components/parameters/RecordId' }]
    post:
      operationId: startOpportunityCheck
      summary: Start an explicit detail check for one selected role
      description: Selected or blocked roles only; selection alone never starts a check. Idempotent on requestKey.
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/CheckStartRequest' } } }
      responses:
        '200':
          description: Exact replay
          content:
            { application/json: { schema: { $ref: '#/components/schemas/CheckStatusView' } } }
        '201':
          description: Check started
          content:
            { application/json: { schema: { $ref: '#/components/schemas/CheckStatusView' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /opportunities/{id}/checks/current:
    parameters: [{ $ref: '#/components/parameters/RecordId' }]
    get:
      operationId: getCurrentOpportunityCheck
      summary: Read the current saved check view for one selected role
      description: Saved proposal with staleness; never triggers work. Side-effect-free.
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Check status
          content:
            { application/json: { schema: { $ref: '#/components/schemas/CheckStatusView' } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /opportunities/{id}/checks/current/activity:
    parameters: [{ $ref: '#/components/parameters/RecordId' }]
    get:
      operationId: listOpportunityCheckActivity
      summary: Read per-role check activity for one selected role
      description: Cursor-paginated role-scoped Codex activity. Side-effect-free.
      security: [{ cookieAuth: [] }]
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200':
          description: Check activity
          content:
            { application/json: { schema: { $ref: '#/components/schemas/CheckActivityPage' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /opportunities/{id}/checks/{checkId}:
    parameters:
      - { $ref: '#/components/parameters/RecordId' }
      - name: checkId
        in: path
        required: true
        schema: { type: string, minLength: 1 }
    get:
      operationId: getOpportunityCheck
      summary: Read one immutable check by id
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Saved check
          content: { application/json: { schema: { $ref: '#/components/schemas/CheckView' } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /answers:
    get:
      operationId: listSavedAnswers
      summary: List owner-approved reusable answers
      security: [{ cookieAuth: [] }]
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100 }
        - name: scope
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Answer library
          content:
            { application/json: { schema: { $ref: '#/components/schemas/SavedAnswerList' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '503': { $ref: '#/components/responses/Unavailable' }
    post:
      operationId: createSavedAnswer
      summary: Create and explicitly approve a reusable answer
      description: Creation confers approved status; unapproved drafts are out of scope.
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/SavedAnswerCreate' } } }
      responses:
        '201':
          description: Answer approved
          content: { application/json: { schema: { $ref: '#/components/schemas/SavedAnswer' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /answers/{answerId}:
    parameters:
      - name: answerId
        in: path
        required: true
        schema: { type: string, minLength: 1 }
    get:
      operationId: getSavedAnswer
      summary: Read one approved answer with versions and provenance
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Saved answer
          content: { application/json: { schema: { $ref: '#/components/schemas/SavedAnswer' } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /answers/{answerId}/versions:
    parameters:
      - name: answerId
        in: path
        required: true
        schema: { type: string, minLength: 1 }
    post:
      operationId: approveSavedAnswerVersion
      summary: Approve a new exact-text version of a saved answer
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          {
            application/json: { schema: { $ref: '#/components/schemas/SavedAnswerVersionCreate' } },
          }
      responses:
        '201':
          description: Version approved
          content: { application/json: { schema: { $ref: '#/components/schemas/SavedAnswer' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /opportunities/{id}/answers/current:
    parameters: [{ $ref: '#/components/parameters/RecordId' }]
    get:
      operationId: getCurrentQuestionAnswers
      summary: Read per-question answer values for the Answer stage
      description: Exact owner values or blanks with provenance. Side-effect-free.
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Answer values
          content:
            { application/json: { schema: { $ref: '#/components/schemas/QuestionAnswerList' } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /opportunities/{id}/questions/{questionId}/answer:
    parameters:
      - { $ref: '#/components/parameters/RecordId' }
      - name: questionId
        in: path
        required: true
        schema: { type: string, minLength: 1 }
    put:
      operationId: saveQuestionAnswer
      summary: Save the owner's exact answer text or an explicit blank
      description: Exact round-trip, no LLM call; blank is a value distinct from unset.
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AnswerValueSave' } } }
      responses:
        '200':
          description: Answer saved
          content:
            { application/json: { schema: { $ref: '#/components/schemas/QuestionAnswerValue' } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  parameters:
    RecordId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jobseek_session
      description: Owner session. Cookie mutations also require X-CSRF-Token and same-origin Origin.
    bearerAuth:
      type: http
      scheme: bearer
      description: Named agent token with operation-specific scopes.
  responses:
    Unavailable:
      description: Required round execution capability is unavailable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    ValidationError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    Conflict:
      description: Revision conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
...
    ResearchActivityEvent:
      type: object
      required: [eventId, at, kind, summary]
      properties:
        eventId: { type: string }
        cursor: { type: string }
        at: { type: string, format: date-time }
        kind: { type: string }
        phase: { type: string, enum: [collect, classify] }
        summary: { type: string }
        refs:
          type: object
          additionalProperties: false
          properties:
            observationId: { type: string }
            captureId: { type: string }
            recordId: { type: string }
            assessmentId: { type: string }
...
    CheckSourceSpan:
      type: object
      required: [captureId, start, end]
      properties:
        captureId: { type: string }
        start: { type: integer, minimum: 0 }
        end: { type: integer, minimum: 0 }
    CheckQuestion:
      type: object
      required: [id, checkId, ordinal, text, required, sourceSpan, sourceExcerpt, textSha256]
      properties:
        id: { type: string }
        checkId: { type: string }
        ordinal: { type: integer, minimum: 0 }
        text: { type: string }
        required: { type: string, enum: [required, optional, unknown] }
        kind: { type: string, enum: [free_text, choice, attachment, other] }
        sourceSpan: { $ref: '#/components/schemas/CheckSourceSpan' }
        sourceExcerpt: { type: string }
        textSha256: { type: string }
    RequestedDocument:
      type: object
      required: [label, required, sourceExcerpt, sourceSpan]
      properties:
        label: { type: string }
        required: { type: boolean }
        sourceExcerpt: { type: string }
        sourceSpan: { $ref: '#/components/schemas/CheckSourceSpan' }
    CheckRoute:
      type: object
      required: [judgment, sourceExcerpt, observedAt]
      properties:
        routeId: { type: string }
        kind: { type: string, enum: [direct, referral, recruiter, unsupported] }
        destinationText: { type: string }
        judgment: { type: string, enum: [application_route, other_contact, unresolved] }
        judgmentSha256: { type: string }
        jevAttemptId: { type: string }
        assessmentId: { type: string }
        sourceExcerpt: { type: string }
        observedAt: { type: string, format: date-time }
    CheckGap:
      type: object
      required: [id, description, consequential, kind]
      properties:
        id: { type: string }
        description: { type: string }
        consequential: { type: boolean }
        kind: { type: string, enum: [missing_fact, ambiguous_source, unverified_claim, other] }
    CheckVacancy:
      type: object
      required: [captureIds, evidenceSourceIds, completeness, sourceUrl, retrievedAt]
      properties:
        captureIds: { type: array, items: { type: string } }
        evidenceSourceIds: { type: array, items: { type: string } }
        completeness: { type: string, enum: [complete, truncated, paginated, partial] }
        sourceUrl: { type: string }
        retrievedAt: { type: string, format: date-time }
    CheckBlockedReason:
      type: object
      required: [code, detail]
      properties:
        code:
          type: string
          enum:
            [source_unavailable, route_ambiguous, route_unsupported, questions_unresolved, other]
        detail: { type: string }
    CheckView:
      type: object
      required:
        [
          id,
          opportunityId,
          opportunityRevision,
          workflowRevision,
          status,
          vacancy,
          requestedDocuments,
          route,
          gaps,
          questions,
          questionSetSha256,
          questionSetVersion,
          createdAt,
          createdBy,
        ]
      properties:
        id: { type: string }
        opportunityId: { type: string }
        opportunityRevision: { type: integer, format: int64 }
        workflowRevision: { type: integer, format: int64 }
        status: { type: string, enum: [checking, checked, blocked] }
        blockedReason: { $ref: '#/components/schemas/CheckBlockedReason' }
        vacancy: { $ref: '#/components/schemas/CheckVacancy' }
        requestedDocuments:
          type: array
          items: { $ref: '#/components/schemas/RequestedDocument' }
        route: { $ref: '#/components/schemas/CheckRoute' }
        gaps: { type: array, items: { $ref: '#/components/schemas/CheckGap' } }
        questions: { type: array, items: { $ref: '#/components/schemas/CheckQuestion' } }
        questionSetSha256: { type: string }
        questionSetVersion: { type: integer, format: int64 }
        activityCursor: { type: string }
        createdAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        createdBy:
          type: object
          required: [actorKind, actorId]
          properties:
            actorKind: { type: string }
            actorId: { type: string }
    CheckStatusView:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [not_checked, checking, checked, blocked, outdated] }
        check: { $ref: '#/components/schemas/CheckView' }
    CheckStartRequest:
      type: object
      additionalProperties: false
      required: [requestKey, expectedOpportunityRevision, expectedWorkflowRevision]
      properties:
        requestKey: { type: string, minLength: 1, maxLength: 200 }
        expectedOpportunityRevision: { type: integer, format: int64, minimum: 1 }
        expectedWorkflowRevision: { type: integer, format: int64, minimum: 0 }
    CheckActivityPage:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items: { $ref: '#/components/schemas/ResearchActivityEvent' }
        nextCursor: { type: string }
    SavedAnswerSourceRef:
      type: object
      required: [kind, ref]
      properties:
        kind: { type: string, enum: [owner_statement, capture, cv_source] }
        ref: { type: string }
        excerpt: { type: string }
    SavedAnswerVersion:
      type: object
      required: [version, text, textSha256, approvedAt, approvedBy, approvalRequestKey]
      properties:
        version: { type: integer, format: int64 }
        text: { type: string }
        textSha256: { type: string }
        approvedAt: { type: string, format: date-time }
        approvedBy:
          type: object
          required: [actorKind, actorId]
          properties:
            actorKind: { type: string }
            actorId: { type: string }
        approvalRequestKey: { type: string }
        changeNote: { type: string }
        sourceRefs:
          type: array
          items: { $ref: '#/components/schemas/SavedAnswerSourceRef' }
        supersedes: { type: integer, format: int64 }
    SavedAnswer:
      type: object
      required: [id, currentVersion, scopeTags, versions]
      properties:
        id: { type: string }
        currentVersion: { type: integer, format: int64 }
        scopeTags: { type: array, items: { type: string } }
        contextNote: { type: string }
        versions:
          type: array
          items: { $ref: '#/components/schemas/SavedAnswerVersion' }
    SavedAnswerList:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/SavedAnswer' } }
        nextCursor: { type: string }
    SavedAnswerCreate:
      type: object
      additionalProperties: false
      required: [requestKey, text, scopeTags]
      properties:
        requestKey: { type: string, minLength: 1, maxLength: 200 }
        text: { type: string, minLength: 1, maxLength: 20000 }
        scopeTags: { type: array, items: { type: string } }
        contextNote: { type: string, maxLength: 2000 }
        sourceRefs:
          type: array
          items: { $ref: '#/components/schemas/SavedAnswerSourceRef' }
    SavedAnswerVersionCreate:
      type: object
      additionalProperties: false
      required: [requestKey, expectedVersion, text]
      properties:
        requestKey: { type: string, minLength: 1, maxLength: 200 }
        expectedVersion: { type: integer, format: int64, minimum: 1 }
        text: { type: string, minLength: 1, maxLength: 20000 }
        changeNote: { type: string, maxLength: 2000 }
    AnswerMatchChoice:
      type: object
      properties:
        answerId: { type: string }
        answerVersion: { type: integer, format: int64 }
        textSha256: { type: string }
        noneFits: { type: boolean }
...
    QuestionAnswerValue:
      type: object
      required:
        [questionId, questionTextSha256, required, version, state, text, provenance, updatedAt]
      properties:
        questionId: { type: string }
        questionTextSha256: { type: string }
        required: { type: string, enum: [required, optional, unknown] }
        version: { type: integer, format: int64 }
        state: { type: string, enum: [unset, answered, blank] }
        text: { type: string }
        textSha256: { type: string }
        provenance:
          type: object
          required: [origin, editedAt, editedBy]
          properties:
            origin:
              type: string
              enum: [jev_suggestion, owner_written, owner_edited, carried_blank]
            matchChoice: { $ref: '#/components/schemas/AnswerMatchChoice' }
            matchId: { type: string }
            editedAt: { type: string, format: date-time }
            editedBy:
              type: object
              required: [actorKind, actorId]
              properties:
                actorKind: { type: string }
                actorId: { type: string }
        updatedAt: { type: string, format: date-time }
    QuestionAnswerList:
      type: object
      required: [checkId, questionSetSha256, values]
      properties:
        checkId: { type: string }
        questionSetSha256: { type: string }
        values:
          type: array
          items: { $ref: '#/components/schemas/QuestionAnswerValue' }
    AnswerValueSave:
      type: object
      additionalProperties: false
      required: [expectedAnswerVersion, text]
      properties:
        expectedAnswerVersion: { type: integer, format: int64, minimum: 0 }
        text: { type: string, maxLength: 20000 }
...